Resolve - change file name contract_addresses to contract-addresses - move get contract address to contract helper
