+++ .gitignore
@@ -4,7 +4,7 @@ coverage
 coverage.json
 .coverage.txt
 .env
-contract_addresses.json
+contract-addresses.json
 
 
 # IntelliJ project files
